add image fitting store
