+++ src/stores/index.ts
@@ -29,3 +29,4 @@ export * from "./ProfileFittingStore";
 export * from "./SnippetStore";
 export * from "./CatalogOnlineQueryProfileStore";
 export * from "./CatalogOnlieQueryConfigStore";
+export * from "./ImageFittingStore";
